+++ src/controllers/logging/ChannelLoggingModel.hpp
@@ -16,11 +16,11 @@ public:
 
 protected:
     // turn a vector item into a model row
-    virtual ChannelLog getItemFromRow(std::vector<QStandardItem *> &row,
+    ChannelLog getItemFromRow(std::vector<QStandardItem *> &row,
                                       const ChannelLog &channelLog) override;
 
     // turns a row in the model into a vector item
-    virtual void getRowFromItem(const ChannelLog &item,
+    void getRowFromItem(const ChannelLog &item,
                                 std::vector<QStandardItem *> &row) override;
 };